fix: move scoreLabel to shared lib for server component access
scoreLabel was in RoastResults.tsx ('use client') but called from
generateMetadata (server). Moved to src/lib/score.ts (no directive).
Re-exported from RoastResults for backward compat.
